// src/types/user.ts
export namespace User {
    export interface ResUser {
        account: string;
        name: string;
        password: string;
        sex: string;
        qq: string;
        email: string;
        phone: string;
        birthday: string;
        hometown: string;
        autograph: string;
        logo: string;
        background: number
        organize_id: number
        state: number;
        create_time: string;
        update_time: string;
        background_url: string;
    }
}